@@ -1,7 +1,7 @@
 package it.k8s.restms.repository;
 
-import it.k8s.restms.model.Item;
+import it.k8s.restms.entity.ItemEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<ItemEntity, Integer> {
 }
